move scrollable emoji block to a separate file
