+++ main.py
@@ -9,6 +9,7 @@ import ctypes
 import pygetwindow
 import pyautogui
 from TitleBar import TitleBar
+from EmojiGrid import EmojiGrid
 from const import *
 
 
@@ -17,63 +18,6 @@ emoji_db = emojis.db.get_emoji_aliases()
 with open("index.css", "r") as f:
 	styles = f.read()
 	f.close()
-
-
-def get_clip():
-	"""
-	gets user's clipboard
-	"""
-	return pyperclip.paste()
-
-
-def add_to_clipboard(txt):
-	"""
-	puts [txt] to the user's clipboard
-
-	:param txt: Text to be copied
-	:return: [txt]
-	"""
-	pyperclip.copy(txt)
-
-	return txt
-
-
-def paste():
-	"""
-	pastes anything from the clipboard
-	"""
-	# i don't like using pyautogui here since it could be ~affected by already
-	# pressed keys
-	pyautogui.hotkey('ctrl', 'v')
-
-
-def activate():
-	"""
-	activates the window behind the EmojiBoard
-
-	:return: Active window's title
-	"""
-	active = [win for win in pygetwindow.getAllWindows() if win.title != ""][1]
-	active.activate()
-
-	return active.title
-
-
-def insert(txt):
-	"""
-	takes the user's original clipboard, changes it to the [txt], pastes it to
-	the active window and retrieves the original clipboard.
-
-	:param txt: Text to be copied/pasted
-	:return: Original text
-	"""
-	original = get_clip()
-	add_to_clipboard(txt)
-	activate()
-	paste()
-	add_to_clipboard(original)
-
-	return txt
 
 
 def main():
@@ -107,51 +51,8 @@ def main():
 
 	tray_icon.setContextMenu(menu)
 
-	# add scrollable area
-	scroll = QScrollArea(parent=window)
-	scroll.setGeometry(10, TITLE_BAR_HEIGHT - 10,
-			WIDTH - 20, HEIGHT - TITLE_BAR_HEIGHT)
-	scroll.setFixedWidth(WIDTH - 20)
-	scroll.setMinimumHeight(HEIGHT - TITLE_BAR_HEIGHT)
-	scroll.setWidgetResizable(True)
-
-	# content for scrollable
-	scroll_contents = QWidget()
-	scroll_contents.setFixedWidth(WIDTH - 40)
-	scroll.setWidget(scroll_contents)
-
-	# using grid layout for scrollable
-	scroll_grid = QGridLayout()
-	scroll_grid.setHorizontalSpacing(4)
-	scroll_grid.setVerticalSpacing(10)
-	scroll_contents.setLayout(scroll_grid)
-
-	i = 0
-
-	# the functionality
-	# thanks stackoverflow
-	font = QFont()
-	font.setFamily("Times")
-	font.setPixelSize(20)
-	for emoji_name, value in emoji_db.items():
-		btn = QPushButton(value, parent=scroll_contents)
-		btn.setToolTip(emoji_name)
-
-		btn.setFont(font)
-		btn.setFixedWidth(42)
-		btn.setFixedHeight(42)
-
-		btn.clicked.connect(lambda ch, txt=value: insert(txt))
-		# btn.clicked.connect(lambda ch, txt=value: add_to_clipboard(txt))
-
-		size_policy = QSizePolicy(QSizePolicy.Maximum,
-				QSizePolicy.Fixed)
-		size_policy.setHorizontalStretch(0)
-		size_policy.setVerticalStretch(0)
-		btn.setSizePolicy(size_policy)
-
-		scroll_grid.addWidget(btn, i // 6, i % 6)
-		i += 1
+	# add scrollable emojis block
+	emoji_grid = EmojiGrid(parent=window)
 
 	# showing stuff
 	tray_icon.show()
